requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
Deprecated==1.2.18
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
Pygments==2.19.2
PyJWT==2.10.1
python-magic==0.4.27
rich==13.9.4
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
wrapt==1.17.2
